chore: Task 5 — move CatalogBrowser out of dead compare/ route, delete vestigial redirect page
- src/app/trips/[tripId]/compare/CatalogBrowser.tsx
    → src/app/trips/[tripId]/components/CatalogBrowser.tsx
- src/app/trips/[tripId]/compare/page.tsx (10-line redirect to /trips/[id]) deleted
- src/app/trips/[tripId]/compare/ directory deleted
- IdeaZonePanel import path updated to './CatalogBrowser'
- Removed stale 'TODO: Move this file' header comment from CatalogBrowser

ArchivedIdeasBrowser's only CatalogBrowser reference was a comment
('Column measurement — mirrors CatalogBrowser') — no import change
needed there.
