--- src/app/trips/[tripId]/compare/page.tsx
@@ -1,10 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default async function ComparePage({
-  params,
-}: {
-  params: Promise<{ tripId: string }>;
-}) {
-  const { tripId } = await params;
-  redirect(`/trips/${tripId}`);
-}
