+++ src/app/trips/[tripId]/components/CatalogBrowser.tsx
@@ -1,4 +1,3 @@
-// TODO: Move this file to src/app/trips/[tripId]/components/ when compare/ directory is cleaned up.
 "use client";
 
 import { useEffect, useLayoutEffect, useRef, useState } from "react";
